chore: remove moduleScopes support in doc
