+++ packages/document/docs/en/config/options/source.mdx
@@ -48,12 +48,6 @@ import GlobalVars from '@en/shared/config/source/globalVars.md';
 
 <GlobalVars />
 
-## source.moduleScopes
-
-import ModuleScopes from '@en/shared/config/source/moduleScopes.md';
-
-<ModuleScopes />
-
 ## source.transformImport
 
 import TransformImport from '@en/shared/config/source/transformImport.md';
