+++ packages/document/docs/en/guide/advanced/rspack-start.mdx
@@ -18,11 +18,7 @@ Currently, most of configuration options in Rsbuild have been adapted for Rspack
 
 #### Source Config
 
-> Source code parsing and compilation related configurations in Rsbuild.
-
-Unsupported configurations and capabilities include:
-
-- [source.moduleScopes](/config/options/source.html#sourcemodulescopes) ([track issue](https://github.com/web-infra-dev/rspack/issues/3548))
+All configurations and capabilities under html are available within rspack.
 
 #### Html Config
