+++ packages/document/docs/zh/config/options/source.mdx
@@ -48,12 +48,6 @@ import GlobalVars from '@zh/shared/config/source/globalVars.md';
 
 <GlobalVars />
 
-## source.moduleScopes
-
-import ModuleScopes from '@zh/shared/config/source/moduleScopes.md';
-
-<ModuleScopes />
-
 ## source.transformImport
 
 import TransformImport from '@zh/shared/config/source/transformImport.md';
