+++ packages/document/docs/zh/guide/advanced/rspack-start.mdx
@@ -20,9 +20,7 @@ Rsbuild 旨在消除不同打包工具之间的主要差异，帮助用户以较
 
 > Rsbuild 中与源代码解析、编译方式相关的配置。
 
-不支持的配置项及能力包括：
-
-- [source.moduleScopes](/config/options/source.html#sourcemodulescopes) ([issue 追踪](https://github.com/web-infra-dev/rspack/issues/3548))
+所有 source 下的配置项及能力在 Rspack 内均可使用。
 
 #### Html Config
