+++ packages/document/docs/zh/guide/start/features.mdx
@@ -14,11 +14,10 @@
 | SourceMap 生成  | 默认在生产环境构建时生成高质量 SourceMap                         | <ul><li>[output.disableSourceMap](/config/options/output.html#outputdisablesourcemap)</li></ul>                                                                                                                   |
 | 文件别名        | 可选功能，通过 alias 设置文件别名                                | <ul><li>[路径别名](/guide/advanced/alias.html)</li><li>[source.alias](/config/options/source.html#sourcealias)</li></ul>                                                                                          |
 | Babel 编译      | 可选功能，通过 SWC 对 JavaScript 和 TypeScript 代码进行转译      | <ul><li>[tools.babel](/config/options/tools.html#toolsbabel)</li></ul>                                                                                                                                            |
-| 限制源代码路径  | 可选功能，限制源代码的引用路径                                   | <ul><li>[source.moduleScopes](/config/options/source.html#sourcemodulescopes)</li></ul>                                                                                                                           |
 | Node 产物       | 可选功能，支持构建出运行在 Node.js 环境的 JavaScript 产物        | <ul><li>[Node 产物](/api/start/build-target.html#node-产物)</li></ul>                                                                                                                                             |
 | Web Worker 产物 | 可选功能，支持构建出运行在 Web Worker 环境的 JavaScript 产物     | <ul><li>[Web Worker 产物](/api/start/build-target.html#web-worker-产物)</li></ul>                                                                                                                                 |
 | 浏览器范围      | 可选功能，通过 browserslist 来设置 Web 应用需要兼容的浏览器范围  | <ul><li>[浏览器范围](/guide/advanced/browserslist.html)</li></ul>                                                                                                                                                 |
-| 兼容性检查      | 可选功能，分析构建产物中是否存在当前浏览器范围下不兼容的高级语法 | <ul><li>[Check Syntax 插件](/plugins/list/plugin-check-syntax.html)</li></ul>                                                                                                                       |
+| 兼容性检查      | 可选功能，分析构建产物中是否存在当前浏览器范围下不兼容的高级语法 | <ul><li>[Check Syntax 插件](/plugins/list/plugin-check-syntax.html)</li></ul>                                                                                                                                     |
 | 注入环境变量    | 可选功能，向代码中注入环境变量或表达式                           | <ul><li>[环境变量](/guide/advanced/define.html)</li></ul>                                                                                                                                                         |
 | Node polyfill   | 可选功能，在浏览器端注入 Node 核心模块的 polyfills               | <ul><li>[Node Polyfill 插件](/plugins/list/plugin-node-polyfill.html)</li></ul>                                                                                                                                   |
 | TS 类型检查     | 可选功能，运行 type checker 检查代码中的类型问题                 | <ul><li>[类型检查](/guide/basic/typescript.html#类型检查)</li></ul>                                                                                                                                               |
